+++ src/Toolkit/kits/shadcn/dialog/assets/controllers/dialog_controller.js
@@ -3,6 +3,16 @@ import { Controller } from '@hotwired/stimulus';
 export default class extends Controller {
     static targets = ['trigger', 'dialog'];
 
+    static values = {
+        open: Boolean,
+    };
+
+    connect() {
+        if (this.openValue) {
+            this.open();
+        }
+    }
+
     open() {
         this.dialogTarget.showModal();